# Reduce computing overhead in SingleTableRule